Use string resources for spinners Breaks EditMoodEventFragment
